requirements.txt
Pillow>=10.0.0
PyPDF4>=1.27.0
PyPDF2>=3.0.0
wand>=0.6.10
tqdm>=4.66.0
natsort>=8.4.0
Flask>=3.0.0
-e .